+++ src/actions-factory.ts
@@ -1,13 +1,12 @@
 import { z } from "zod";
 import { Action } from "./action";
 import { Config } from "./config";
-import { EmissionMap } from "./emission";
 import { ActionContext, Handler } from "./handler";
-import { Namespaces, RootNS } from "./namespaces";
+import { Namespaces, RootNS } from "./namespace";
 
 interface Commons<
   IN extends z.AnyZodTuple,
-  NS extends Namespaces<EmissionMap>,
+  NS extends Namespaces,
   K extends keyof NS,
 > {
   /** @desc The incoming event payload validation schema (without or excluding acknowledgement) */
@@ -23,35 +22,41 @@ interface Commons<
 
 export interface ActionNoAckDef<
   IN extends z.AnyZodTuple,
-  NS extends Namespaces<EmissionMap>,
+  NS extends Namespaces,
   K extends keyof NS,
 > extends Commons<IN, NS, K> {
   /** @desc No output schema => no returns => no acknowledgement */
-  handler: Handler<ActionContext<z.output<IN>, NS[K]>, void>;
+  handler: Handler<
+    ActionContext<z.output<IN>, NS[K]["emission"], NS[K]["metadata"]>,
+    void
+  >;
 }
 
 export interface ActionWithAckDef<
   IN extends z.AnyZodTuple,
   OUT extends z.AnyZodTuple,
-  NS extends Namespaces<EmissionMap>,
+  NS extends Namespaces,
   K extends keyof NS,
 > extends Commons<IN, NS, K> {
   /** @desc The acknowledgement validation schema */
   output: OUT;
   /** @desc The returns become an Acknowledgement */
-  handler: Handler<ActionContext<z.output<IN>, NS[K]>, z.input<OUT>>;
+  handler: Handler<
+    ActionContext<z.output<IN>, NS[K]["emission"], NS[K]["metadata"]>,
+    z.input<OUT>
+  >;
 }
 
-export class ActionsFactory<NS extends Namespaces<EmissionMap>> {
+export class ActionsFactory<NS extends Namespaces> {
   constructor(protected config: Config<NS>) {}
 
   public build<
     IN extends z.AnyZodTuple,
     OUT extends z.AnyZodTuple,
-    K extends keyof NS & string = RootNS,
+    K extends keyof NS = RootNS,
   >(
     def: ActionNoAckDef<IN, NS, K> | ActionWithAckDef<IN, OUT, NS, K>,
-  ): Action<IN, OUT> {
+  ): Action<IN, OUT, NS> {
     return new Action(def);
   }
 }