+++ src/attach.spec.ts
@@ -1,7 +1,9 @@
 import http from "node:http";
 import { Server } from "socket.io";
 import { describe, expect, test, vi } from "vitest";
+import { z } from "zod";
 import { attachSockets } from "./attach";
+import { createConfig } from "./config";
 import { AbstractLogger } from "./logger";
 
 describe("Attach", () => {
@@ -65,13 +67,18 @@ describe("Attach", () => {
         io: ioMock as unknown as Server,
         target: targetMock as unknown as http.Server,
         actions: actionsMock,
-        config: {
+        config: createConfig({
           startupLogo: false,
           timeout: 100,
-          emission: { "/": {} },
+          namespaces: {
+            "/": {
+              emission: {},
+              hooks,
+              metadata: z.object({ name: z.string() }),
+            },
+          },
           logger: loggerMock as unknown as AbstractLogger,
-        },
-        hooks,
+        }),
       });
       expect(ioMock.of).toHaveBeenLastCalledWith("/");
       expect(ioMock.attach).toHaveBeenCalledWith(targetMock);
@@ -175,6 +182,11 @@ describe("Attach", () => {
       actionsMock[0].execute.mock.lastCall[0].client.setData({
         name: "user",
       });
+      expect(() =>
+        actionsMock[0].execute.mock.lastCall[0].client.setData({
+          name: 123,
+        }),
+      ).toThrow(z.ZodError);
 
       // client.getData:
       expect(actionsMock[0].execute.mock.lastCall[0].client.getData()).toEqual({