+++ src/attach.ts
@@ -4,34 +4,17 @@ import { AbstractAction } from "./action";
 import { Client } from "./client";
 import { Config } from "./config";
 import { makeDistribution } from "./distribution";
-import {
-  EmissionMap,
-  EmitterConfig,
-  makeEmitter,
-  makeRoomService,
-} from "./emission";
+import { EmitterConfig, makeEmitter, makeRoomService } from "./emission";
 import { ClientContext, IndependentContext } from "./handler";
-import { HookSet, Hooks } from "./hooks";
-import {
-  Namespaces,
-  RootNS,
-  ensureNamespaces,
-  normalizeNS,
-} from "./namespaces";
+import { Namespaces, normalizeNS } from "./namespace";
 import { getRemoteClients } from "./remote-client";
 import { getStartupLogo } from "./startup-logo";
 
-export const attachSockets = async <NS extends Namespaces<EmissionMap>>({
+export const attachSockets = async <NS extends Namespaces>({
   io,
   actions,
   target,
-  config: {
-    logger: rootLogger,
-    emission: namespaces,
-    timeout,
-    startupLogo = true,
-  },
-  hooks: hooksCfg,
+  config: { logger: rootLogger, namespaces, timeout, startupLogo = true },
 }: {
   /**
    * @desc The Socket.IO server
@@ -50,17 +33,12 @@ export const attachSockets = async <NS extends Namespaces<EmissionMap>>({
   target: http.Server;
   /** @desc The configuration describing the emission (outgoing events) */
   config: Config<NS>;
-  hooks?: Hooks<NS> | HookSet<NS[RootNS]>;
 }): Promise<Server> => {
-  const hooks = ensureNamespaces(
-    hooksCfg || {},
-    (value) => typeof value === "function",
-  );
-
   for (const name in namespaces) {
-    type NSEmissions = NS[typeof name];
+    type NSEmissions = NS[typeof name]["emission"];
+    type NSMeta = NS[typeof name]["metadata"];
     const ns = io.of(normalizeNS(name));
-    const emission = namespaces[name];
+    const { emission, hooks, metadata } = namespaces[name];
     const {
       onConnection = ({ client: { id, getData }, logger }) =>
         logger.debug("Client connected", { ...getData(), id }),
@@ -71,11 +49,11 @@ export const attachSockets = async <NS extends Namespaces<EmissionMap>>({
       onAnyOutgoing = ({ event, logger, payload }) =>
         logger.debug(`Sending ${event}`, payload),
       onStartup = ({ logger }) => logger.debug("Ready"),
-    } = (hooks?.[name] || {}) as HookSet<NSEmissions>;
+    } = hooks;
     const emitCfg: EmitterConfig<NSEmissions> = { emission, timeout };
-    const nsCtx: IndependentContext<NSEmissions> = {
+    const nsCtx: IndependentContext<NSEmissions, NSMeta> = {
       logger: rootLogger,
-      withRooms: makeRoomService({ subject: io, ...emitCfg }),
+      withRooms: makeRoomService({ subject: io, metadata, ...emitCfg }),
       all: {
         getClients: async () => getRemoteClients(await ns.fetchSockets()),
         getRooms: () => Array.from(ns.adapter.rooms.keys()),
@@ -85,20 +63,23 @@ export const attachSockets = async <NS extends Namespaces<EmissionMap>>({
     ns.on("connection", async (socket) => {
       const emit = makeEmitter({ subject: socket, ...emitCfg });
       const broadcast = makeEmitter({ subject: socket.broadcast, ...emitCfg });
-      const client: Client<NSEmissions> = {
+      const client: Client<NSEmissions, NSMeta> = {
         emit,
         broadcast,
         id: socket.id,
         isConnected: () => socket.connected,
         getRooms: () => Array.from(socket.rooms),
         getData: () => socket.data || {},
-        setData: (value) => (socket.data = value),
+        setData: (value) => {
+          metadata.parse(value); // validation only, no transformations
+          socket.data = value;
+        },
         ...makeDistribution(socket),
       };
-      const ctx: ClientContext<NSEmissions> = {
+      const ctx: ClientContext<NSEmissions, NSMeta> = {
         ...nsCtx,
         client,
-        withRooms: makeRoomService({ subject: socket, ...emitCfg }),
+        withRooms: makeRoomService({ subject: socket, metadata, ...emitCfg }),
       };
       await onConnection(ctx);
       socket.onAny((event, ...payload) =>