+++ src/hooks.ts
@@ -1,3 +1,4 @@
+import { z } from "zod";
 import { EmissionMap } from "./emission";
 import {
   ClientContext,
@@ -5,18 +6,13 @@ import {
   IndependentContext,
   TracingContext,
 } from "./handler";
-import { Namespaces } from "./namespaces";
 
-export interface HookSet<E extends EmissionMap> {
+export interface Hooks<E extends EmissionMap, D extends z.SomeZodObject> {
   /** @desc A place for emitting events regardless receiving events */
-  onConnection?: Handler<ClientContext<E>, void>;
-  onDisconnect?: Handler<ClientContext<E>, void>;
-  onAnyIncoming?: Handler<TracingContext<E>, void>;
-  onAnyOutgoing?: Handler<TracingContext<E>, void>;
+  onConnection?: Handler<ClientContext<E, D>, void>;
+  onDisconnect?: Handler<ClientContext<E, D>, void>;
+  onAnyIncoming?: Handler<TracingContext<E, D>, void>;
+  onAnyOutgoing?: Handler<TracingContext<E, D>, void>;
   /** @desc A place for emitting events regardless clients activity */
-  onStartup?: Handler<IndependentContext<E>, void>;
+  onStartup?: Handler<IndependentContext<E, D>, void>;
 }
-
-export type Hooks<NS extends Namespaces<EmissionMap>> = {
-  [K in keyof NS]?: HookSet<NS[K]>;
-};