+++ src/integration.ts
@@ -2,7 +2,6 @@ import ts from "typescript";
 import { z } from "zod";
 import { AbstractAction } from "./action";
 import { Config } from "./config";
-import { EmissionMap } from "./emission";
 import {
   defaultSerializer,
   exportModifier,
@@ -10,12 +9,12 @@ import {
   makeCleanId,
   makeEventFnSchema,
 } from "./integration-helpers";
-import { Namespaces, normalizeNS } from "./namespaces";
+import { Namespaces, normalizeNS } from "./namespace";
 import { zodToTs } from "./zts";
 import { addJsDocComment, createTypeAlias, printNode } from "./zts-helpers";
 
 interface IntegrationProps {
-  config: Config<Namespaces<EmissionMap>>;
+  config: Config<Namespaces>;
   actions: AbstractAction[];
   /**
    * @desc When event has both .rest() and an acknowledgement, the "...rest" can not be placed in a middle.
@@ -91,7 +90,7 @@ export class Integration {
   }
 
   constructor({
-    config: { emission: namespaces },
+    config: { namespaces },
     actions,
     serializer = defaultSerializer,
     optionalPropStyle = { withQuestionMark: true, withUndefined: true },
@@ -115,7 +114,7 @@ export class Integration {
       ),
     );
 
-    for (const [ns, emission] of Object.entries(namespaces)) {
+    for (const [ns, { emission }] of Object.entries(namespaces)) {
       this.aliases[ns] = {};
       this.registry[ns] = { emission: [], actions: [] };
       const commons = {