+++ src/remote-client.ts
@@ -1,17 +1,20 @@
 import { RemoteSocket } from "socket.io";
+import { z } from "zod";
 
 import { Distribution, makeDistribution } from "./distribution";
 
-export interface RemoteClient extends Distribution {
+export interface RemoteClient<D extends z.SomeZodObject> extends Distribution {
   id: string;
   rooms: string[];
-  getData: <D extends object>() => Readonly<Partial<D>>;
+  getData: () => Readonly<Partial<z.infer<D>>>;
 }
 
-export const getRemoteClients = (sockets: RemoteSocket<{}, unknown>[]) =>
-  sockets.map<RemoteClient>((socket) => ({
+export const getRemoteClients = <D extends z.SomeZodObject>(
+  sockets: RemoteSocket<{}, z.infer<D>>[],
+) =>
+  sockets.map<RemoteClient<D>>((socket) => ({
     id: socket.id,
     rooms: Array.from(socket.rooms),
-    getData: <D extends object>() => (socket.data as D) || {},
+    getData: () => socket.data || {},
     ...makeDistribution(socket),
   }));